requirements.txt
numpy
sympy
pillow
pillow-avif-plugin
pyperclipimg